+++ Apps/Examples/Examples/All Examples/Annotations/ViewAnnotationBasicExample.swift
@@ -43,34 +43,30 @@ final class ViewAnnotationBasicExample: UIViewController, ExampleProtocol {
     private func addViewAnnotation(at coordinate: CLLocationCoordinate2D) {
         let options = ViewAnnotationOptions(
             geometry: Point(coordinate),
-            allowOverlap: false,
+            allowOverlap: true,
             anchor: .center
         )
-        let labelText = String(format: "lat=%.2f\nlon=%.2f", coordinate.latitude, coordinate.longitude)
-        let sampleView = createSampleView(withText: labelText)
-        try? mapView.viewAnnotations.add(sampleView, options: options)
+        let annotationView = AnnotationView(frame: CGRect(x: 0, y: 0, width: 100, height: 80))
+        annotationView.title = String(format: "lat=%.2f\nlon=%.2f", coordinate.latitude, coordinate.longitude)
+        annotationView.delegate = self
+        try? mapView.viewAnnotations.add(annotationView, options: options)
+    }
+}
+
+extension ViewAnnotationBasicExample: AnnotationViewDelegate {
+    func annotationViewDidSelect(_ annotationView: AnnotationView) {
+        let options = ViewAnnotationOptions(selected: true)
+
+        try? mapView.viewAnnotations.update(annotationView, options: options)
     }
 
-    private func createSampleView(withText text: String) -> UIView {
-        let sampleView = UIView()
-        sampleView.bounds.size = CGSize(width: 64, height: 32)
-        sampleView.backgroundColor = .blue
-        sampleView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(onSampleViewClick)))
+    func annotationViewDidUnselect(_ annotationView: AnnotationView) {
+        let options = ViewAnnotationOptions(selected: false)
 
-        let label = UILabel()
-        label.text = text
-        label.font = UIFont.systemFont(ofSize: 10)
-        label.numberOfLines = 0
-        label.textColor = .white
-        label.translatesAutoresizingMaskIntoConstraints = false
-        sampleView.addSubview(label)
-
-        NSLayoutConstraint.activate([
-            label.centerXAnchor.constraint(equalTo: sampleView.centerXAnchor),
-            label.centerYAnchor.constraint(equalTo: sampleView.centerYAnchor)
-        ])
-
-        return sampleView
+        try? mapView.viewAnnotations.update(annotationView, options: options)
     }
 
+    func annotationViewDidPressClose(_ annotationView: AnnotationView) {
+        mapView.viewAnnotations.remove(annotationView)
+    }
 }