+++ Apps/Examples/Examples/Controllers/ExampleTableViewController.swift
@@ -94,8 +94,16 @@ extension ExampleTableViewController {
           example = examples[indexPath.row]
         }
 
-        let cell = UITableViewCell(style: .default, reuseIdentifier: "reuseIdentifier")
+        let cell = UITableViewCell(style: .subtitle, reuseIdentifier: "reuseIdentifier")
         cell.textLabel?.text = example.title
+        cell.textLabel?.numberOfLines = 2
+        cell.detailTextLabel?.text = example.description.trimmingCharacters(in: .whitespacesAndNewlines)
+        cell.detailTextLabel?.numberOfLines = 2
+        if #available(iOS 13.0, *) {
+            cell.detailTextLabel?.textColor = .secondaryLabel
+        } else {
+            cell.detailTextLabel?.textColor = .lightGray
+        }
         return cell
     }